Put MCPToolsAdapter in separate file and set up to accept list of URLs.
